pod 'NSInteger' should not be used as format arguments; add an explicit cast to 'long' instead [-Wformat]

+++ XLsn0wKit/XLsn0wComponents/XLsn0wCountTimeButton/XLsn0wCountTimeButton.m
@@ -72,7 +72,7 @@
         }
         else
         {
-            NSString *title = [NSString stringWithFormat:@"%zd秒",_second];
+            NSString *title = [NSString stringWithFormat:@"%zd秒",(long)_second];
             [self setTitle:title forState:UIControlStateNormal];
             [self setTitle:title forState:UIControlStateDisabled];
